Removed stdlib.
Removed stdlib
Fixed some crashes
Redid cmake to work with VSCode, and the Bartman Debugger.

Builds and runs fine via VSCode using Bartman's GCC.
Blue screen crash was replicated if I didn't have the font on the disk.
Fixed crash when it tries to draw more than 8 blocks.
Made the font texture bigger as a work around for the score limit or 10000.